requirements.txt
streamlit
pandas
numpy
joblib
matplotlib
seaborn
qrcode
pillow
scikit-learn
xgboost
openpyxl
plotly
